+++ n0core/scheduler/agent/porter.py
@@ -1,7 +1,7 @@
-from scheduler import consumer, logger, send  # NOQA
+from initialize import consumer, logger, send  # NOQA
 try:
     from n0core.lib import proto
 except:  # NOQA
     import sys
     sys.path.append('../../')
-    from n0core.lib import proto
+    from n0core.lib import proto  # NOQA
