+++ n0core/scheduler/scheduler.py
@@ -1,63 +1,4 @@
 import sqlalchemy as sa
-from urllib.parse import urlparse
-from n0library.arguments.common import CommonArguments
-from n0library.logger import Logger
-try:
-    from n0core.lib.n0mq import N0MQ
-except:  # NOQA
-    import sys
-    sys.path.append('../../')
-    from n0core.lib.n0mq import N0MQ
-
-# from ResorceCalculation import Execfunc
-
-
-parser = CommonArguments(
-    description="",
-)
-
-parser.add_argument("--pulsar-url",
-                    type=str,
-                    default='pulsar://127.0.0.1:6650',
-                    dest="pulsar_url",
-                    help="Pulsar URL")
-parser.add_argument("--scheduler-topic",
-                    type=str,
-                    default='persistent://main/sd/scheduler/handle',
-                    dest="scheduler_topic",
-                    help="Scheduler Topic")
-parser.add_argument("--database-url",
-                    type=str,
-                    default=None,
-                    dest="db_url",
-                    help="Database URL")
-parser.add_argument("--prometheus-url",
-                    type=str,
-                    default=None,
-                    dest="prometheus_url",
-                    help="Prometheus URL")
-args = parser.parse_args()
-
-pulsar_url = args.pulsar_url
-scheduler_topic = args.scheduler_topic
-db_url = urlparse(args.db_url)
-prometheus_url = urlparse(args.prometheus_url)
-
-
-client = N0MQ(pulsar_url)
-consumer = client.subscribe(scheduler_topic)
-logger = Logger(name='scheduler', stdout=True, level='info')
-
-
-def send(url, req):
-    producer = client.create_producer(url)
-    producer.send(req)
-
-
-def CheckHost(msg):
-    if msg.host is None:
-        print("host")
-    return send(msg)
 
 
 class Scheduler():
@@ -85,13 +26,6 @@ class Scheduler():
         re.DBResorce(self.DB_URL)
         re.PrometheusRresorce(self.PROMETHEUS_HOST, self.PROMETHEUS_PORT)
         return re.ResorceCalculation(Execfunc)
-
-
-scheduler = Scheduler(
-    db_url=db_url,
-    prometheus_host=prometheus_url.hostname,
-    prometheus_port=str(prometheus_url.port),
-)
 
 
 class Resorce():
